Move the FindPathRequest struct to the PathFinderRequest.h
