+++ gemrb/core/PathFinderRequest.h
@@ -23,12 +23,15 @@
 
 // The vocabulary types of a pathfinding request
 
+#include "Region.h"
+
 #include <cstddef>
 #include <cstdint>
 #include <limits>
 
 namespace GemRB {
 
+class Map;
 class Movable;
 
 /**
@@ -78,6 +81,103 @@ static_assert(static_cast<size_t>(FindPathRequestPriority::Immediate) == Schedul
 	      "ScheduledQueuesPrioritiesCount undercounts the queued priorities and Sync() "
 	      "indexes workerScheduledQueuesByPriority out of bounds");
 
+
+/**
+ * Represents a request to find a path between two locations in the navigation system.
+ * Encapsulates all necessary information for the pathfinding algorithm to compute a route,
+ * including start and end positions, navigation constraints, and request-specific parameters.
+ */
+struct FindPathRequest {
+	FindPathRequestPriority priority = FindPathRequestPriority::Normal;
+	Point source;
+	Point destination;
+	SearchmapPoint actorSMPos; // actor's searchmap position, snapshotted at request time
+	// The Movable that issued this request. Never dereferenced on worker threads: it is used
+	// only for pointer identity comparisons (skipping the requester's own footprint when
+	// repainting the searchmap, and recognising its own cell as bumpable). Actor is the only
+	// Movable subclass, so comparisons against the const Actor* stored in ActorSearchMapData
+	// and TraversabilityCellData are plain upcasts.
+	Movable* instigatorIdentity = nullptr;
+	Map* map = nullptr; // never dereferenced on worker threads, used in immediate calculation flow and for getting ID
+	int actorCircleSize = 0; // actor's collision circle size, snapshotted at request time
+	unsigned int minDistance = 0;
+	int actorSpeed = 0; // actor's walk speed, snapshotted at request time
+	int pathfindingFlags = 0;
+	FindPathRequestType requestType = FindPathRequestType::WalkTo;
+	bool canRePathIgnoringActors = false;
+	bool blocksSearchMaps = false;
+
+	/**
+	 *  Sets the basic request metadata.
+	 *  Builder method for setting request type, priority, and the instigating Movable.
+	 *
+	 *  @param inRequestType The type of pathfinding operation (WalkTo, RunAway, etc.)
+	 *  @param inPriority The scheduling priority for this request
+	 *  @param inInstigatorIdentity Pointer identity of the Movable that initiated the request
+	 *  @return Reference to this request for method chaining
+	 */
+	FindPathRequest& PutBasicRequestData(const FindPathRequestType inRequestType,
+					     const FindPathRequestPriority inPriority,
+					     Movable* inInstigatorIdentity)
+	{
+		requestType = inRequestType;
+		priority = inPriority;
+		instigatorIdentity = inInstigatorIdentity;
+		return *this;
+	}
+
+	/**
+	 *  Sets the pathfinding spatial data.
+	 *  Builder method for setting source, destination, map reference, and pathfinding flags.
+	 *
+	 *  @param inSource The starting position for pathfinding
+	 *  @param inDestination The target position for pathfinding
+	 *  @param inMap Pointer identity of the Map containing the path (not dereferenced on worker threads)
+	 *  @param inPathfindingFlags Bitfield of pathfinding behavior flags
+	 *  @return Reference to this request for method chaining
+	 */
+	FindPathRequest& PutPathData(
+		const Point& inSource,
+		const Point& inDestination,
+		Map* inMap,
+		const int inPathfindingFlags)
+	{
+		source = inSource;
+		destination = inDestination;
+		map = inMap;
+		pathfindingFlags = inPathfindingFlags;
+		return *this;
+	}
+
+	/**
+	 *  Sets the actor-specific pathfinding data.
+	 *  Builder method for setting collision properties, movement parameters,
+	 *  and retry behavior flags. The requester's identity is set by PutBasicRequestData().
+	 *
+	 *  @param inActorSMPos Actor's searchmap position snapshot
+	 *  @param inActorCircleSize Actor's collision circle size snapshot
+	 *  @param inMinDistance Minimum acceptable distance from destination
+	 *  @param inActorSpeed Actor's walk speed snapshot
+	 *  @param inCanRePathIgnoringActors Whether pathfinding should retry ignoring actors if it fails
+	 *  @param inBlocksSearchMaps Whether this actor blocks searchmaps
+	 *  @return Reference to this request for method chaining
+	 */
+	FindPathRequest& PutActorData(const SearchmapPoint& inActorSMPos,
+				      const int inActorCircleSize,
+				      const unsigned int inMinDistance,
+				      const int inActorSpeed,
+				      const bool inCanRePathIgnoringActors,
+				      const bool inBlocksSearchMaps)
+	{
+		actorSMPos = inActorSMPos;
+		actorCircleSize = inActorCircleSize;
+		minDistance = inMinDistance;
+		actorSpeed = inActorSpeed;
+		canRePathIgnoringActors = inCanRePathIgnoringActors;
+		blocksSearchMaps = inBlocksSearchMaps;
+		return *this;
+	}
+};
 
 /**
  *  Unique identifier for pathfinding requests in the navigation system.
